Fix HashedWheelTimer when timeout is multiple of wheel period #2898
